requirements.txt
Aspose.Email-for-Python-via-NET